+++ tests/utils/workflow.py
@@ -6,6 +6,11 @@ from typing import MutableSequence, cast, TYPE_CHECKING
 from streamflow.core import utils
 from streamflow.core.deployment import Target
 from streamflow.core.config import BindingConfig
+from streamflow.workflow.combinator import (
+    DotProductCombinator,
+    CartesianProductCombinator,
+    LoopTerminationCombinator,
+)
 from streamflow.workflow.port import ConnectorPort
 from streamflow.core.workflow import Workflow, Port
 from streamflow.workflow.step import DeployStep, ScheduleStep
@@ -27,7 +32,12 @@ def create_deploy_step(workflow, deployment_config=None):
     )
 
 
-def create_schedule_step(workflow, deploy_step, binding_config=None):
+def create_schedule_step(
+    workflow: Workflow,
+    deploy_steps: MutableSequence[DeployStep],
+    binding_config: BindingConfig = None,
+):
+    # It is necessary to pass in the correct order biding_config.targets and deploy_steps for the mapping
     if not binding_config:
         binding_config = BindingConfig(
             targets=[
@@ -35,6 +45,7 @@ def create_schedule_step(workflow, deploy_step, binding_config=None):
                     deployment=deploy_step.deployment_config,
                     workdir=utils.random_name(),
                 )
+                for deploy_step in deploy_steps
             ]
         )
     return workflow.create_step(
@@ -42,7 +53,8 @@ def create_schedule_step(workflow, deploy_step, binding_config=None):
         name=posixpath.join(utils.random_name(), "__schedule__"),
         job_prefix="something",
         connector_ports={
-            binding_config.targets[0].deployment.name: deploy_step.get_output_port()
+            target.deployment.name: deploy_step.get_output_port()
+            for target, deploy_step in zip(binding_config.targets, deploy_steps)
         },
         binding_config=binding_config,
     )
@@ -59,3 +71,28 @@ async def create_workflow(
         ports.append(workflow.create_port())
     await workflow.save(context)
     return workflow, tuple(cast(MutableSequence[Port], ports))
+
+
+def get_dot_combinator():
+    return DotProductCombinator(name=utils.random_name(), workflow=None)
+
+
+def get_cartesian_product_combinator():
+    return CartesianProductCombinator(name=utils.random_name(), workflow=None)
+
+
+def get_loop_terminator_combinator():
+    c = LoopTerminationCombinator(name=utils.random_name(), workflow=None)
+    c.add_output_item("test1")
+    c.add_output_item("test2")
+    return c
+
+
+def get_nested_crossproduct():
+    combinator = DotProductCombinator(name=utils.random_name(), workflow=None)
+    c1 = CartesianProductCombinator(name=utils.random_name(), workflow=None)
+    c1.add_item("ext")
+    c1.add_item("inn")
+    items = c1.get_items(False)
+    combinator.add_combinator(c1, items)
+    return combinator
